ciuGUI—a simple tool for comparing two CIU datasets
Two .csv files with the standard CIU format are imported--these must
contain the same number of rows and columns. The data can then be
smoothed with either 1D area or intensity normalisation.

The data are displayed as a four-pane figure. The top two panes show the
heat-maps for each data set in the selected colormap. The bottom-left
pane shows the RMSD heat-map and the rolling RMSD (across the collision
energies). The bottom-right pane shows the overlaid mobiligrams for a
selected energy, the corresponding energy is also displayed as a line on
the top panes.